+++ app/routes/data-api/project/citizen-scientist/pattern-matchings.js
@@ -8,22 +8,37 @@ var model = require('../../../../model');
 var APIError = require('../../../../utils/apierror');
 var csv_stringify = require("csv-stringify");
 
-
-
-/** Return a list of all the pattern matchings in a project.
+/**
+ * Return a list of all the pattern matchings in a project.
  */
 router.get('/', function(req, res, next) {
     res.type('json');
-    model.patternMatchings.find({
+    const opts = {
         project:req.project.project_id,
-        citizen_scientist:1,
-        deleted:0,
+        citizen_scientist: 1,
+        deleted: 0,
+    }
+    // get stats for validated vs total rois first
+    model.patternMatchings.find({
+        ...opts,
         showUserStatsFor: req.session.user.id,
-        showSpecies: true,
-        showTemplate: true,
-        showPlaylist:true
-    }).then(function(count) {
-        res.json(count);
+        userStats: ['validated'],
+        showTemplate: true
+    }).then(async function(pms1) {
+        // then get additional data about species and songtype
+        const pms2 = await model.patternMatchings.find({
+            ...opts,
+            showSpecies: true,
+        })
+        // match first set of data with second set
+        const result = pms1.map((pm1) => {
+            const additionalPm = pms2.find(pm2 => pm2.id === pm1.id)
+            return {
+                ...pm1,
+                ...additionalPm
+            }
+        })
+        res.json(result);
     }).catch(next);
 });